+++ fj-doc-playground-quarkus/src/main/java/org/fugerit/java/doc/playground/doc/GenerateRest.java
@@ -1,50 +1,23 @@
 package org.fugerit.java.doc.playground.doc;
 
 import java.io.ByteArrayOutputStream;
-import java.io.Reader;
 import java.io.StringReader;
-import java.io.StringWriter;
 import java.util.Base64;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
 
-import org.fugerit.java.core.function.SafeFunction;
-import org.fugerit.java.core.io.StreamIO;
 import org.fugerit.java.core.lang.helpers.BooleanUtils;
 import org.fugerit.java.core.util.checkpoint.CheckpointUtils;
 import org.fugerit.java.core.util.result.Result;
-import org.fugerit.java.doc.base.config.DocInput;
-import org.fugerit.java.doc.base.config.DocOutput;
 import org.fugerit.java.doc.base.config.DocTypeHandler;
-import org.fugerit.java.doc.base.config.DocTypeHandlerXMLUTF8;
-import org.fugerit.java.doc.base.facade.DocFacadeSource;
 import org.fugerit.java.doc.base.parser.DocParser;
 import org.fugerit.java.doc.base.parser.DocValidationResult;
-import org.fugerit.java.doc.base.typehandler.markdown.SimpleMarkdownExtTypeHandler;
-import org.fugerit.java.doc.freemarker.config.FreeMarkerConfigStep;
-import org.fugerit.java.doc.freemarker.fun.SimpleMessageFun;
-import org.fugerit.java.doc.freemarker.html.FreeMarkerHtmlFragmentTypeHandlerEscapeUTF8;
 import org.fugerit.java.doc.lib.simpletable.SimpleTableDocConfig;
 import org.fugerit.java.doc.lib.simpletable.SimpleTableFacade;
 import org.fugerit.java.doc.lib.simpletable.model.SimpleRow;
 import org.fugerit.java.doc.lib.simpletable.model.SimpleTable;
-import org.fugerit.java.doc.mod.poi.XlsxPoiTypeHandler;
 import org.fugerit.java.doc.playground.RestHelper;
-import org.fugerit.java.doc.playground.config.InitPlayground;
-import org.fugerit.java.doc.playground.facade.BasicInput;
-import org.fugerit.java.doc.playground.facade.InputFacade;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-import freemarker.cache.StringTemplateLoader;
-import freemarker.template.Configuration;
-import freemarker.template.Template;
-import freemarker.template.TemplateExceptionHandler;
-import freemarker.template.Version;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
@@ -58,117 +31,8 @@ import lombok.extern.slf4j.Slf4j;
 @Path("/generate")
 public class GenerateRest {
 
-	public static final String FTL_DIRECTIVE = "<#ftl";
-	
-	private void doHandle( DocTypeHandler handler, String type, int sourceType, Reader reader, ByteArrayOutputStream baos ) {
-		SafeFunction.apply( () -> {
-			DocInput docInput = DocInput.newInput( type, reader , sourceType );
-			DocOutput docOutput = DocOutput.newOutput( baos );
-			handler.handle(docInput, docOutput);
-		} );
-	}
-	
-	private void handleConfiguration( Configuration configuration, JsonNode node, String ftlData, String chainId ) {
-		StringTemplateLoader loader = new StringTemplateLoader();
-		StringBuilder chainData = new StringBuilder();
-		ObjectNode oNode = (ObjectNode) node;
-		Iterator<String> fieldNames = oNode.fieldNames();
-		int ftlDirectiveStartIndex = ftlData.indexOf( FTL_DIRECTIVE );
-		if ( ftlDirectiveStartIndex != -1 ) {
-			int ftlDirectiveEndIndex = ftlData.indexOf( ">" );
-			String ftlDirective = ftlData.substring( 0, ftlDirectiveEndIndex+1 );
-			log.debug( "ftlDirective : {}", ftlDirective );
-			chainData.append( ftlDirective );
-			ftlData = ftlData.substring( ftlDirectiveEndIndex+1 );
-		}
-		while ( fieldNames.hasNext() ) {
-			String currentName = fieldNames.next();
-			String currentValue = node.get( currentName ).toString();
-			log.debug( "handleConfiguration() set var {} -> {}", currentName, currentValue );
-			chainData.append( "<#assign " );
-			chainData.append( currentName );
-			chainData.append( "=" );
-			chainData.append( currentValue );
-			chainData.append( ">");
-		}
-		chainData.append( ftlData );
-		loader.putTemplate( chainId , chainData.toString() );
-		configuration.setTemplateExceptionHandler( TemplateExceptionHandler.RETHROW_HANDLER );
-		configuration.setTemplateLoader( loader );
-	}
-	
-	private void handleFtlx( DocTypeHandler handler, String type, int sourceType, Reader reader, ByteArrayOutputStream baos, String freemarkerJsonData ) {
-		SafeFunction.apply( () -> {
-			// volatile FreeMarker Template configuration
-			String chainId = "current_"+System.currentTimeMillis();
-			Configuration configuration = new Configuration( new Version( FreeMarkerConfigStep.ATT_FREEMARKER_CONFIG_KEY_VERSION_LATEST ) );
-			ObjectMapper mapper = new ObjectMapper();
-			try ( StringReader jsonReader = new StringReader(freemarkerJsonData) ) {
-				JsonNode node = mapper.readTree( jsonReader ); // parse json node to sanitize input
-				this.handleConfiguration(configuration, node, StreamIO.readString( reader ), chainId );
-				Template currentChain = configuration.getTemplate( chainId );
-				Map<Object, Object> data = new HashMap<>();
-				data.put( "messageFormat" , new SimpleMessageFun() );
-				try ( StringWriter writer = new StringWriter() ) {
-					currentChain.process( data , writer );
-					try ( StringReader ftlReader = new StringReader( writer.toString() ) ) {
-						this.doHandle(handler, type, sourceType, ftlReader, baos);
-					}
-				}
-			}
-			configuration.clearTemplateCache();
-		} );
-	}
-	
-	private byte[] generateHelper( GenerateInput input, DocTypeHandler handler) throws Exception {
-		byte[] result = null;
-		if ( input.getDocContent() != null ) {
-			try ( StringReader reader = new StringReader( input.getDocContent() );
-					ByteArrayOutputStream baos = new ByteArrayOutputStream() ) {
-				int sourceType = DocFacadeSource.SOURCE_TYPE_XML;
-				if ( InputFacade.FORMAT_JSON.equalsIgnoreCase( input.getInputFormat() ) ) {
-					sourceType = DocFacadeSource.SOURCE_TYPE_JSON;
-				} else if ( InputFacade.FORMAT_YAML.equalsIgnoreCase( input.getInputFormat() ) ) {
-					sourceType = DocFacadeSource.SOURCE_TYPE_YAML;
-				}
-				String type = input.getOutputFormat().toLowerCase();
-				log.info( "output format : {}", type );
-				if ( InputFacade.FORMAT_FTLX.equalsIgnoreCase( input.getInputFormat() ) ) {
-					this.handleFtlx(handler, type, sourceType, reader, baos, input.getFreemarkerJsonData());
-				} else {
-					this.doHandle(handler, type, sourceType, reader, baos);
-				}
-				result =  baos.toByteArray();
-			}
-		}
-		return result;
-	}
-
-	private DocTypeHandler findHandler( BasicInput input ) {
-		DocTypeHandler handler = InitPlayground.PDF_FOP_TYPE_HANDLER;
-		if ( "XLSX".equalsIgnoreCase( input.getOutputFormat() ) ) {
-			handler = XlsxPoiTypeHandler.HANDLER;
-		} else if ( "HTML".equalsIgnoreCase( input.getOutputFormat() ) ) {
-			handler = FreeMarkerHtmlFragmentTypeHandlerEscapeUTF8.HANDLER;
-		} else if ( "PDFA".equalsIgnoreCase( input.getOutputFormat() ) ) {
-			handler = InitPlayground.PDFA_FOP_TYPE_HANDLER;
-		} else if ( "XML".equalsIgnoreCase( input.getOutputFormat() ) ) {
-			handler = DocTypeHandlerXMLUTF8.HANDLER;
-		} else if ( "MD".equalsIgnoreCase( input.getOutputFormat() ) ) {
-			handler = SimpleMarkdownExtTypeHandler.HANDLER_NOCOMMENTS_UTF8;
-		}
-		return handler;
-	}
-	
-	private DocParser findParser( BasicInput input ) {
-		int sourceType = DocFacadeSource.SOURCE_TYPE_XML;
-		if ( InputFacade.FORMAT_JSON.equalsIgnoreCase( input.getInputFormat() ) ) {
-			sourceType = DocFacadeSource.SOURCE_TYPE_JSON;
-		} else if ( InputFacade.FORMAT_YAML.equalsIgnoreCase( input.getInputFormat() ) ) {
-			sourceType = DocFacadeSource.SOURCE_TYPE_YAML;
-		}
-		return DocFacadeSource.getInstance().getParserForSource( sourceType );
-	}
+	@Inject
+	private GenerateFacade facade;
 	
 	private Throwable findCause( Throwable o ) {
 		return o.getCause() != null ? this.findCause( o.getCause() ) : o;
@@ -181,10 +45,10 @@ public class GenerateRest {
 	public Response document( GenerateInput input) {
 		return RestHelper.defaultHandle( () -> {
 			long time = System.currentTimeMillis();
-			DocTypeHandler handler = this.findHandler(input);
+			DocTypeHandler handler = facade.findHandler(input);
 			GenerateOutput output = new GenerateOutput();
 			try {
-				byte[] data = this.generateHelper(input, handler);
+				byte[] data = facade.generateHelper(input, handler);
 				output.setDocOutputBase64( Base64.getEncoder().encodeToString( data ) );
 				output.setGenerationTime( CheckpointUtils.formatTimeDiffMillis( time , System.currentTimeMillis() ) );
 			} catch ( Exception e ) {
@@ -214,11 +78,11 @@ public class GenerateRest {
 	@Path("/validate")
 	public Response validate( GenerateInput input) {
 		return RestHelper.defaultHandle( () -> {
-			DocParser parser = this.findParser(input);
+			DocParser parser = facade.findParser(input);
 			try ( StringReader reader = new StringReader( input.getDocContent() );
 					ByteArrayOutputStream buffer = new ByteArrayOutputStream() ) {
 				DocValidationResult result = parser.validateResult( reader );
-				DocTypeHandler handler = this.findHandler(input);
+				DocTypeHandler handler = facade.findHandler(input);
 				SimpleTable simpleTableModel = SimpleTableFacade.newTable( 15, 20, 65 );
 		        SimpleRow headerRow = new SimpleRow( BooleanUtils.BOOLEAN_TRUE );
 		        headerRow.addCell( "#" );